requirements.txt
requests==2.22.0
discord.py==1.5.0
discord==1.0.1
PyYAML==5.3.1
